move Authentication to shardingsphere-common
