+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/sharding/core/yaml/swapper/AuthenticationYamlSwapper.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.sharding.core.yaml.swapper;
 
 import com.google.common.collect.Maps;
-import org.apache.shardingsphere.sharding.core.rule.Authentication;
+import org.apache.shardingsphere.underlying.common.auth.Authentication;
 import org.apache.shardingsphere.sharding.core.yaml.config.common.YamlAuthenticationConfiguration;
 import org.apache.shardingsphere.underlying.common.yaml.swapper.YamlSwapper;
