+++ sharding-core/sharding-core-common/src/test/java/org/apache/shardingsphere/sharding/core/yaml/swapper/AuthenticationYamlSwapperTest.java
@@ -17,8 +17,8 @@
 
 package org.apache.shardingsphere.sharding.core.yaml.swapper;
 
-import org.apache.shardingsphere.sharding.core.rule.Authentication;
-import org.apache.shardingsphere.sharding.core.rule.ProxyUser;
+import org.apache.shardingsphere.underlying.common.auth.Authentication;
+import org.apache.shardingsphere.underlying.common.auth.ProxyUser;
 import org.apache.shardingsphere.sharding.core.yaml.config.common.YamlAuthenticationConfiguration;
 import org.apache.shardingsphere.sharding.core.yaml.config.common.YamlProxyUserConfiguration;
 import org.junit.Test;
